add mock query and mock mutation #78
