+++ zenenti-app/mock/mock.js
@@ -1,0 +1,88 @@
+const mock = {
+  user: {
+    name: "Zeny",
+  },
+  practice: {
+    categories: ["Calma", "Dormir", "Bienestar", "Enfoque"],
+    types: ["Respiración", "Meditación"],
+    practices: [
+      {
+        category: "Bienestar",
+        type: ["Meditación"],
+        name: "Intención Positiva",
+        durationMillis: 104000,
+        level: 1,
+        audio: require("../assets/audio/practice/BI_ME_1-5_N1_Intención positiva.m4a"),
+      },
+      {
+        category: "Calma",
+        type: ["Meditación"],
+        name: "Meditación guiada de un minuto para la calma",
+        durationMillis: 61000,
+        level: 1,
+        audio: require("../assets/audio/practice/CA_ME_1-5_N1_Meditación guiada de un minuto para la calma.mp4"),
+      },
+    ],
+  },
+  quotes: ["«Dar no nos empobrece, ni retener nos enriquece.» B.K.S"],
+};
+
+/**
+ * Utility function for querying into the mock data structure. Allows you to make 'fetch'-like requests with a query string.
+ * @param queryString Path to item.
+ * @returns Query function.
+ */
+const mockQuery = (queryString) => {
+  return async () => {
+    const pathElements = queryString.split("/");
+
+    let queryItem = mock;
+    for (const pathElement of pathElements) {
+      if (queryItem[pathElement]) {
+        queryItem = queryItem[pathElement];
+        continue;
+      }
+
+      if (Array.isArray(queryItem)) {
+        queryItem = queryItem[Number(pathElement)] || null;
+        continue;
+      }
+
+      throw `${pathElement} does not exist.`;
+    }
+
+    return queryItem;
+  };
+};
+
+/**
+ * Function to generate a mutation against the mock data.
+ * @param queryString
+ * @returns {(function(*): Promise<void>)|*}
+ */
+const mockMutation = (queryString) => {
+  return async (value) => {
+    const pathElements = queryString.split("/");
+
+    let queryItem = mock;
+    for (let i = 0; i < pathElements.length - 1; i++) {
+      const pathElement = pathElements[i];
+
+      if (queryItem[pathElement]) {
+        queryItem = queryItem[pathElement];
+        continue;
+      }
+
+      if (Array.isArray(queryItem)) {
+        queryItem = queryItem[Number(pathElement)] || null;
+        continue;
+      }
+
+      throw `${pathElement} does not exist.`;
+    }
+
+    queryItem[pathElements[pathElements.length - 1]] = value;
+  };
+};
+
+export { mockQuery, mockMutation };
